perf(ammalloc): optimize memory layout and refactor benchmarks
Memory layout optimizations:
- FreeList: change size_/max_size_ from size_t to uint32_t
  (reduces FreeList from 24B to 16B, better cache locality)
- SpanList: add alignas(CACHE_LINE_SIZE) to prevent false sharing
- ThreadCache: use SystemConfig::CACHE_LINE_SIZE instead of hardcoded 64

SizeClass optimizations:
- Add AM_ALWAYS_INLINE to Index(), Size(), RoundUp()
- Ensures hot path functions are inlined for maximum performance

ThreadCache simplification:
- Remove ReleaseTooLongList(), merge logic into DeallocateSlowPath()
- Cleaner code structure

Benchmark refactoring:
- Use sliding window pattern (BM_Malloc_Churn) for memory control
- Test scenarios: Window=1 (fast path), 256 (steady state), 1024 (overflow)
- Add BM_Malloc_Deep_Churn for CentralCache stress test
- Random size test uses window to prevent memory explosion

+++ include/ammalloc/central_cache.h
@@ -8,6 +8,7 @@
 #include "ammalloc/size_class.h"
 #include "ammalloc/span.h"
 
+#include <cstdint>
 #include <mutex>
 
 namespace aethermind {
@@ -76,8 +77,8 @@ public:
 
 private:
     FreeBlock* head_;
-    size_t size_;
-    size_t max_size_;
+    uint32_t size_;
+    uint32_t max_size_;
 };
